fix(cli): move firewall remediation rendering into ChalkConsoleLogger
The FIREWALL_BLOCKED_NON_INTERACTIVE handler in the run command wrote
chalk-colored console.warn lines unconditionally, regardless of the
selected logger. Under `--logger json` this injected colored text into
the middle of the JSONL stream, corrupting it. The JSON logger already
emits this event as a structured line, so the extra rendering was both
redundant there and harmful to machine-readable output.

Move the rendering into ChalkConsoleLogger as a handleFirewallBlocked
handler (subscribed/disposed like the other handlers), so colored
remediation output only appears with the chalk logger and the JSON
logger emits a clean JSON line. Drop the run-command subscription and
the now-unused printFirewallRemediation export.

The two remediation-rendering tests move from the CLI run command tests
to the core logger tests, exercising the handler through the event
emitter.
