+++ packages/cli/test/commands/run.test.ts
@@ -1,7 +1,6 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import { Command } from "commander";
-import { createRunCommand, printFirewallRemediation } from "../../src/commands/run.js";
-import type { FirewallBlockedNonInteractiveEventData } from "pilo-core";
+import { createRunCommand } from "../../src/commands/run.js";
 import { getConfigDefaults } from "pilo-core";
 
 // Get defaults from schema (used for mocking config.getConfig)
@@ -587,88 +586,5 @@ describe("CLI Run Command", () => {
       // Just check that the process exited with error code
       expect(mockExit).toHaveBeenCalledWith(1);
     });
-  });
-});
-
-describe("printFirewallRemediation", () => {
-  let warnSpy: ReturnType<typeof vi.spyOn>;
-
-  beforeEach(() => {
-    warnSpy = vi.spyOn(console, "warn").mockImplementation(() => {});
-  });
-
-  afterEach(() => {
-    warnSpy.mockRestore();
-  });
-
-  it("prints all three remediation options with the blocked hostname", () => {
-    const data: FirewallBlockedNonInteractiveEventData = {
-      timestamp: Date.now(),
-      iterationId: "",
-      reason: "Security policy blocked submitting a form containing unauthorized agent-filled data",
-      kind: "form-submission",
-      pageHostname: "untrusted.com",
-      formActionHostnames: ["untrusted.com"],
-      remediations: [
-        {
-          kind: "add-trusted-hostnames",
-          hostnames: ["untrusted.com"],
-          description: "Add untrusted.com to trusted_hostnames to allow this action on this site.",
-        },
-        {
-          kind: "enable-interactive-mode",
-          description:
-            "Run in interactive mode by providing a UserDataCallback so the agent can ask the user to approve sensitive fields per-action via request_user_data.",
-        },
-        {
-          kind: "enable-unsafe-mode",
-          description: "Set unsafe_mode=true to disable the action firewall entirely. WARNING: ...",
-        },
-      ],
-    };
-
-    printFirewallRemediation(data);
-    const output = warnSpy.mock.calls
-      .map((c: unknown[]) => c.join(" "))
-      .join("\n")
-      .replace(/\x1b\[[0-9;]*m/g, "");
-    expect(output).toContain("untrusted.com");
-    expect(output).toContain("trusted_hostnames untrusted.com");
-    expect(output).toContain("interactive mode");
-    expect(output).toContain("unsafe_mode true");
-  });
-
-  it("falls back to a generic command when no hostnames are listed", () => {
-    const data: FirewallBlockedNonInteractiveEventData = {
-      timestamp: Date.now(),
-      iterationId: "",
-      reason: "Security policy blocked filling a submittable form field without user approval",
-      kind: "freeform-fill",
-      pageHostname: null,
-      formActionHostnames: [],
-      remediations: [
-        {
-          kind: "add-trusted-hostnames",
-          hostnames: [],
-          description:
-            "Add the page hostname to trusted_hostnames to allow this action on this site.",
-        },
-        {
-          kind: "enable-interactive-mode",
-          description: "Run in interactive mode...",
-        },
-        {
-          kind: "enable-unsafe-mode",
-          description: "Set unsafe_mode=true...",
-        },
-      ],
-    };
-
-    printFirewallRemediation(data);
-    const output = warnSpy.mock.calls
-      .map((c: unknown[]) => c.join(" "))
-      .join("\n")
-      .replace(/\x1b\[[0-9;]*m/g, "");
-    expect(output).toContain("trusted_hostnames <host>");
   });
 });
