+++ packages/core/test/loggers.test.ts
@@ -1,6 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import { ConsoleLogger } from "../src/loggers/console.js";
 import { JSONConsoleLogger } from "../src/loggers/json.js";
+import { ChalkConsoleLogger } from "../src/loggers/chalkConsole.js";
 import { WebAgentEventEmitter, WebAgentEventType } from "../src/events.js";
 import type {
   TaskStartEventData,
@@ -21,6 +22,7 @@ import type {
   ProcessingEventData,
   ScreenshotCapturedEventData,
   ScreenshotCapturedImageEventData,
+  FirewallBlockedNonInteractiveEventData,
 } from "../src/events.js";
 
 // Mock console methods
@@ -1074,6 +1076,108 @@ describe("JSONConsoleLogger", () => {
       const output = mockConsole.log.mock.calls[0][0];
       const parsed = JSON.parse(output);
       expect(parsed.event).toBe(WebAgentEventType.BROWSER_SCREENSHOT_CAPTURED);
+    });
+  });
+});
+
+describe("ChalkConsoleLogger", () => {
+  let logger: ChalkConsoleLogger;
+  let emitter: WebAgentEventEmitter;
+  let warnSpy: ReturnType<typeof vi.spyOn>;
+
+  beforeEach(() => {
+    warnSpy = vi.spyOn(console, "warn").mockImplementation(() => {});
+    logger = new ChalkConsoleLogger();
+    emitter = new WebAgentEventEmitter();
+    logger.initialize(emitter);
+  });
+
+  afterEach(() => {
+    logger.dispose();
+    warnSpy.mockRestore();
+  });
+
+  // Strip ANSI color codes so assertions can match plain text.
+  const plainOutput = () =>
+    warnSpy.mock.calls
+      .map((c: unknown[]) => c.join(" "))
+      .join("\n")
+      .replace(/\x1b\[[0-9;]*m/g, "");
+
+  describe("firewall remediation", () => {
+    it("prints all three remediation options with the blocked hostname", () => {
+      const data: FirewallBlockedNonInteractiveEventData = {
+        timestamp: Date.now(),
+        iterationId: "",
+        reason:
+          "Security policy blocked submitting a form containing unauthorized agent-filled data",
+        kind: "form-submission",
+        pageHostname: "untrusted.com",
+        formActionHostnames: ["untrusted.com"],
+        remediations: [
+          {
+            kind: "add-trusted-hostnames",
+            hostnames: ["untrusted.com"],
+            description:
+              "Add untrusted.com to trusted_hostnames to allow this action on this site.",
+          },
+          {
+            kind: "enable-interactive-mode",
+            description:
+              "Run in interactive mode by providing a UserDataCallback so the agent can ask the user to approve sensitive fields per-action via request_user_data.",
+          },
+          {
+            kind: "enable-unsafe-mode",
+            description:
+              "Set unsafe_mode=true to disable the action firewall entirely. WARNING: ...",
+          },
+        ],
+      };
+
+      emitter.emitEvent({
+        type: WebAgentEventType.FIREWALL_BLOCKED_NON_INTERACTIVE,
+        data,
+      });
+
+      const output = plainOutput();
+      expect(output).toContain("untrusted.com");
+      expect(output).toContain("trusted_hostnames untrusted.com");
+      expect(output).toContain("interactive mode");
+      expect(output).toContain("unsafe_mode true");
+    });
+
+    it("falls back to a generic command when no hostnames are listed", () => {
+      const data: FirewallBlockedNonInteractiveEventData = {
+        timestamp: Date.now(),
+        iterationId: "",
+        reason: "Security policy blocked filling a submittable form field without user approval",
+        kind: "freeform-fill",
+        pageHostname: null,
+        formActionHostnames: [],
+        remediations: [
+          {
+            kind: "add-trusted-hostnames",
+            hostnames: [],
+            description:
+              "Add the page hostname to trusted_hostnames to allow this action on this site.",
+          },
+          {
+            kind: "enable-interactive-mode",
+            description: "Run in interactive mode...",
+          },
+          {
+            kind: "enable-unsafe-mode",
+            description: "Set unsafe_mode=true...",
+          },
+        ],
+      };
+
+      emitter.emitEvent({
+        type: WebAgentEventType.FIREWALL_BLOCKED_NON_INTERACTIVE,
+        data,
+      });
+
+      expect(plainOutput()).toContain("trusted_hostnames <host>");
     });
   });
 });
